% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_patch.R
\name{geom_patch}
\alias{geom_patch}
\title{Title}
\usage{
geom_patch(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  linejoin = "mitre",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{inherit.aes}{}
}
\description{
Title
}
