% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pattern.R
\name{geom_pattern}
\alias{geom_pattern}
\title{Title}
\usage{
geom_pattern(
  mapping = NULL,
  data = NULL,
  stat = "count",
  position = "stack",
  ...,
  just = 0.5,
  width = NULL,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{inherit.aes}{}
}
\description{
Title
}
