% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_symbol.R
\name{geom_symbol}
\alias{geom_symbol}
\title{Title}
\usage{
geom_symbol(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{inherit.aes}{}
}
\description{
Title
}
