% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_symbolswarm.R
\name{geom_symbolswarm}
\alias{geom_symbolswarm}
\title{Title}
\usage{
geom_symbolswarm(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  ...,
  method = "swarm",
  cex = 1,
  side = 0L,
  priority = "ascending",
  fast = TRUE,
  dodge.width = NULL,
  corral = "none",
  corral.width = 0.9,
  groupOnX = NULL,
  beeswarmArgs = list(),
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{inherit.aes}{}
}
\description{
Title
}
