% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchGrob.R
\name{patchGrob}
\alias{patchGrob}
\title{Title}
\usage{
patchGrob(
  xmin = unit(0.2, "npc"),
  xmax = unit(0.8, "npc"),
  ymin = unit(0.2, "npc"),
  ymax = unit(0.8, "npc"),
  pattern = "e_hori",
  spacing = unit(2, "mm"),
  p_linewidth = 0.5,
  p_colour = "black",
  fill = "black",
  name = NULL,
  gp = NULL,
  vp = NULL
)
}
\arguments{
\item{vp}{}
}
\description{
Title
}
