% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_patch.R
\name{scale_p_linewidth_manual}
\alias{scale_p_linewidth_manual}
\title{Title}
\usage{
scale_p_linewidth_manual(..., values, breaks = waiver(), na.value = NA)
}
\arguments{
\item{na.value}{}
}
\description{
Title
}
